Feat: add useFetchSingleVideo
